# [TOPI] improve inclusive_scan for thrust