The `topi.cuda.inclusive_scan` currently relies on performing an `exclusive_scan` followed by an add operation that adds the input data back in. To eliminate the overhead introduced by this extra addition, probably we should have an implementation specifically designed for `inclusive_scan`. As of now, the implementation of `inclusive_scan` is limited to `exclusive_scan_ir`, developing an `inclusive_scan_ir` will be needed for performance in the future.

In this pull request, we specifically address this efficiency for thrust by directly calling the inclusive function.

Co-authored-by: Wuwei Lin  [wuwei@apache.org](mailto:wuwei@apache.org)